Proyecto-2-BACKEND/requirements.txt
asgiref==3.8.1
automata-lib==8.4.0
cached_method==0.1.0
contourpy==1.3.1
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Django==5.2
djangorestframework==3.16.0
fonttools==4.55.0
frozendict==2.4.6
kiwisolver==1.4.7
matplotlib==3.9.2
mutagen==1.47.0
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pygame==2.6.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.9.0
django-filter==23.2
google-auth==2.24.0
google-auth-oauthlib==1.0.0
google-api-python-client==2.91.0
cryptography==41.0.4
